# src/book/models.py
# generated by fastapi-codegen:
#   filename:  openapi/specification/book-webapp-fastapi_API-specification.yaml
#   timestamp: 2022-09-06T17:09:01+00:00

from __future__ import annotations

from datetime import date, datetime
from typing import Optional

from pydantic import BaseModel, Field


class Book(BaseModel):
    id: str = Field(..., description='Unique identifier for the given book.')
    title: str
    author: Optional[str] = None
    category: str
    country: Optional[str] = None
    publishDate: Optional[date] = Field(
        None,
        description="Set to true if the book's email has been verified.",
        example='2022-08-10',
    )
    createdDatetime: Optional[datetime] = Field(
        None, description='The date that the book was created.'
    )
    updatedDatetime: Optional[datetime] = Field(
        None, description='The date that the book was created.'
    )


class BookRequest(BaseModel):
    id: str = Field(..., description='Unique identifier for the given book.')
    title: str
    author: Optional[str] = None
    category: str
    country: Optional[str] = None
    publishDate: Optional[date] = Field(
        None,
        description="Set to true if the book's email has been verified.",
        example='2022-08-10',
    )


class BooksBookIdGetResponse(BaseModel):
    pass


class BooksBookIdGetResponse1(BaseModel):
    pass


class BooksBookIdGetResponse2(BaseModel):
    pass


class BooksBookIdPutResponse(BaseModel):
    pass


class BooksBookIdDeleteResponse(BaseModel):
    pass


class BooksPostResponse(BaseModel):
    pass